[APAM-564] filter logic for mit & cit constents and disable mit consent deletion (#265)
* [APAM-532] schema viewmodel, some independent new classes, and duplicate (for now) some functions to Airwallex.kt
- those functions shouldn't belong to an activity anymore, it should be a shared function
- fix small typo on analytics
- created centralized list of Analytics names. new ones should use this but for now we're not migrating the existing
- irrelevant change in AirwallexActivity but it's to fix window leak issue

* [APAM-532] paymentOperationsViewModel only

* [APAM-532] wrapper for payment element

* [APAM-532] context instead of activity to support fragment & composable

* [APAM-532] use sharedflow for result

* [APAM-532] Card-related changes to use PaymentElement & listener

* [APAM-532] payment methods integration

* [APAM-532] schema payment listener and ensure padding is handled by client/activity

* [APAM-532] optimize imports

* [APAM-532] fix add new card shouldn't check for AirwallexPaymentSession

* [APAM-532] some formatting fixes for gradle detekt

* [APAM-532] fix unit test

* [APAM-532] more code cleanup

* [APAM-532] add more unit tests

* [APAM-532] fix detekt issues

* [APAM-532] fix unit test

* [APAM-532] more unit tests

* [APAM-532] small adjustments to name

* [APAM-532] updated code for detekt

* [APAM-532] re-rename layoutType

* [APAM-532] fix potential mockk leak

* [APAM-532] remove airwallex.kt exclusion

* [APAM-531] base code for support for custom theme and dark mode

* [APAM-531] small cleanup and sample theme color

* [APAM-532] rename operation to flow

* [APAM-532] rename according to previous MR change

* [APAM-531] default theme will be airwallex_tint_color and can be overriden

* [APAM-531] implement new color schemes to existing views

* [APAM-531] support theme color change for xml files

* [APAM-531] change the way to set theme by configuration paymentAppearance

* [APAM-531] disable payment appearance customization in sample app

* [APAM-534] embedded element sample

* [APAM-532] rename for ios consistency

* Revert "[APAM-532] rename for ios consistency"

This reverts commit c0cbe5be65dcdc81ad34502ec9a5a2cb1ca57689.

* [APAM-532] remove paymentMethodsViewModel

* [APAM-532] rename payment element for ios consistency

* [APAM-531] revert xml styles, manually set colors instead to support force theme mode

* [APAM-531] set base background for the whole component

* [APAM-534] support dark & light theme

* [APAM-531] reflection function to support appearance change using embedded without AirwallexStarter

* [APAM-552] more analytics on session

* [APAM-531] fix detekt

* [APAM-552] internal function for payment element create for analytics

* [APAM-531] fix unit test

* [APAM-531] fix detekt

* [APAM-552] unit tests

* [APAM-531] reflection must always be called regardless paymentAppearance exist or not

* [APAM-530] Support displaying Google Pay as part of the method tab/accordion

* [APAM-551] schema loading is handled locally

* [APAM-551] default loading dialog

* [APAM-530] change icon

* [APAM-534] gpay options for embedded element

* [APAM-555] add maestro tests for embedded element

* [APAM-552] fix payment_launched trigger, and remove subtype since we have launchType

* [APAM-555] remove hardcoded cust id

* [APAM-552] remove println

* [APAM-532] rename variable name

* [APAM-532] refactor for smaller composable

* [APAM-551] use dialog fragment instead for lifecycle awareness

* [APAM-563] add Session and modify intent confirm request to support simplifying flow

* [APAM-564] filter logic for mit & cit constents and disable mit consent deletion

* [APAM-532] handle condition where only gpay in the list

* [APAM-532] never fetch methods & consents if config is Card

* [APAM-532] fix logic & change exception type

* [APAM-564] add paymentIntentProvider in Session and change to safer param types for consent options

* [APAM-564] remove init block

* [APAM-564] add paymentIntentProvider in Session and change to safer param types for consent options

* [APAM-563] fix payment schedule

* [APAM-564] fix for detekt

* [APAM-531] fix code comment

* [APAM-531] AirwallexTheme to use AirwallexColor

* [APAM-531] use variable instead of function for color names

* [APAM-552] missing conflict resolution

* [APAM-534] embedded sample extends basePaymentActivity

* [APAM-534] revamp the sample products to be expandable

* [APAM-534] add payment methods title

* [APAM-552] update analytics

* [APAM-552] move dialog setup session to oncreate

* [APAM-552] fix detekt

* [APAM-530] add prioritizeGooglePay to analytics

* [APAM-551] color name change

* [APAM-561] fix detekt

* [APAM-564] fix unit test

* refactor: add schema viewmodel and shared functions (APAM-532)


* refactor: extract paymentFlowViewModel (APAM-532)


* refactor: add wrapper for payment element (APAM-532)


* feat: add card payment element and listener (APAM-532)


* feat: add payment methods integration (APAM-532)


* fix: gradle detekt and AddCard session check (APAM-532)


* test: add unit tests and code cleanup for detekt (APAM-532)


* feat: add base for dark mode and custom theme (APAM-531)


* feat: add embedded element sample (APAM-534)


* feat: add more analytics on session (APAM-552)


* feat: support displaying Google Pay as part of the method tab/accordion (APAM-530)


* feat: add local loading schema and default loading behavior (APAM-551)


* fix: config change issue during checkout (APAM-551)


* feat: embedded element inline gpay & card only
fix: potential race condition for loading dialog fragment

* feat: update english documentation

* fix: bindToActivity should not be in sample app

* fix: change assertion due to UI change

* fix: supportedCardBrands will have default value

* fix: supportedCardBrands will have default value

* fix: detekt

* chore: change CardScheme to AirwallexSupportedCard in PaymentElement

* chore: updated guide & guide-zh

* feat: java support

* chore: use AnalyticsLogger.Field constants for analytics event field names

- Add Field object in AnalyticsLogger as single source of truth for all event field key constants
- Replace all literal field name strings with Field.* constants across source and test files
- Fix snake_case field name "payment_method" to camelCase "paymentMethod"


* fix: add missing German (de) translation strings


* [APAM-560] feat: Update documentation in guide.md (#274)

* feat: update english documentation

* fix: supportedCardBrands will have default value

* fix: supportedCardBrands will have default value

* chore: change CardScheme to AirwallexSupportedCard in PaymentElement

* chore: updated guide & guide-zh

* feat: java support

* chore: not using PaymentElementHelper class anymore

* fix: remove java activity from android manifest

* refactor: renderIn instance method

* test: unit tests and sonar exclusion


* chore: remove unnecessary java impl code

* chore: remove unused code due to wrong conflict resolution

* test: fix maestro test cases

* fix: embedded element express not working due to null paymentIntentProvider

* fix: check providerId first before paymentIntentProvider before binding activity

* fix: don't return anything if invalid(null) fingerprint consent

* test: fix testcase

---------

+++ airwallex/src/main/java/com/airwallex/android/view/composables/consent/ConsentListSection.kt
@@ -12,7 +12,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.airwallex.android.R
+import com.airwallex.android.core.TokenManager
 import com.airwallex.android.core.model.PaymentConsent
+import com.airwallex.android.ui.composables.AirwallexColor
 import com.airwallex.android.ui.composables.ScreenView
 import com.airwallex.android.ui.composables.StandardAlertDialog
 import java.util.Locale
@@ -45,26 +47,43 @@ internal fun ConsentListSection(
     }
 
     if (localConsentToBeDeleted != null) {
-        StandardAlertDialog(
-            title = stringResource(
-                R.string.airwallex_delete_consent_alert_title,
-                String.format(
-                    "%s •••• %s",
-                    localConsentToBeDeleted?.paymentMethod?.card?.brand?.uppercase(Locale.ROOT)
-                        .orEmpty(),
-                    localConsentToBeDeleted?.paymentMethod?.card?.last4.orEmpty(),
-                ),
-            ),
-            content = stringResource(id = R.string.airwallex_delete_consent_alert_content),
-            confirmButtonTitle = stringResource(R.string.airwallex_delete_payment_method_positive),
-            dismissButtonTitle = stringResource(R.string.airwallex_delete_payment_method_negative),
-            onConfirm = {
-                localConsentToBeDeleted?.let { onDeleteCard(it) }
-                localConsentToBeDeleted = null
-            },
-            onDismiss = {
-                localConsentToBeDeleted = null
-            },
+        val cardDisplay = String.format(
+            "%s •••• %s",
+            localConsentToBeDeleted?.paymentMethod?.card?.brand?.uppercase(Locale.ROOT)
+                .orEmpty(),
+            localConsentToBeDeleted?.paymentMethod?.card?.last4.orEmpty(),
         )
+
+        if (localConsentToBeDeleted?.nextTriggeredBy == PaymentConsent.NextTriggeredBy.MERCHANT) {
+            // Show informational dialog for merchant-triggered consents
+            val businessName = TokenManager.businessName ?: ""
+            StandardAlertDialog(
+                title = stringResource(R.string.airwallex_merchant_consent_info_title),
+                content = stringResource(
+                    id = R.string.airwallex_merchant_consent_info_content,
+                    businessName, businessName
+                ),
+                dismissButtonTitle = stringResource(R.string.airwallex_delete_payment_method_negative),
+                onDismiss = {
+                    localConsentToBeDeleted = null
+                },
+            )
+        } else {
+            // Show delete confirmation dialog for customer-triggered consents
+            StandardAlertDialog(
+                title = stringResource(R.string.airwallex_delete_consent_alert_title, cardDisplay),
+                content = stringResource(id = R.string.airwallex_delete_consent_alert_content),
+                confirmButtonTitle = stringResource(R.string.airwallex_delete_payment_method_positive),
+                dismissButtonTitle = stringResource(R.string.airwallex_delete_payment_method_negative),
+                onConfirm = {
+                    localConsentToBeDeleted?.let { onDeleteCard(it) }
+                    localConsentToBeDeleted = null
+                },
+                onDismiss = {
+                    localConsentToBeDeleted = null
+                },
+                confirmButtonContainerColor = AirwallexColor.textError
+            )
+        }
     }
 }
